Refactor, preparing to run QX in gtest
